requirement.txt
Django>=4.2.7
djangorestframework>=3.14.0
pycardano>=0.9.0
blockfrost-python>=0.6.0
setuptools>=68.2.2